Refactored deals.service.ts to extract constants and interfaces
