+++ GamerpotFrontend/src/library/models/dealsSearchParams.ts
@@ -1,0 +1,9 @@
+export interface DealsSearchParams {
+  [key: string]: any;
+  pageNumber: number;
+  pageSize?: number;
+  params?: string;
+  title?: string;
+  lowerPrice?: string;
+  upperPrice?: string;
+}
