+++ GamerpotFrontend/src/library/services/deals.service.ts
@@ -1,26 +1,11 @@
-import { buildParams } from './paramBuilder';
-import { DealInterface } from '../models/deal';
-
-export interface DealsSearchParams {
-  [key: string]: any;
-  pageNumber: number;
-  pageSize?: number;
-  params?: string;
-  title?: string;
-  lowerPrice?: string;
-  upperPrice?: string;
-}
-
-const DEFAULT_URL_PARAMS: DealsSearchParams = {
-  pageNumber: 0,
-  title: '',
-  pageSize: 15,
-  lowerPrice: '0',
-  upperPrice: '500',
-};
-
-const API_DEALS_URL = 'https://www.cheapshark.com/api/1.0/deals?';
-const API_STORES_URL = 'https://www.cheapshark.com/api/1.0/stores';
+import {buildParams} from './paramBuilder';
+import {DealInterface} from '../models/deal';
+import {DealsSearchParams} from '../models/dealsSearchParams';
+import {
+  DEFAULT_URL_PARAMS,
+  API_DEALS_URL,
+  API_STORES_URL,
+} from '../constants/dealsConstants';
 
 const getStores = async () => {
   const response = await fetch(API_STORES_URL);
@@ -33,12 +18,10 @@ export const getDeals = async (params?: DealsSearchParams) => {
   const deals: DealInterface[] = await response.json();
 
   const stores = await getStores();
-  const fixedDeals: DealInterface[] = deals.map(deal => (
-    {
-      ...deal,
-      ...stores[deal.storeID - 1],
-    }
-  ));
+  const fixedDeals: DealInterface[] = deals.map(deal => ({
+    ...deal,
+    ...stores[deal.storeID - 1],
+  }));
 
   return fixedDeals;
 };
