+++ GamerpotFrontend/src/library/services/paramBuilder.ts
@@ -4,7 +4,7 @@ interface UrlParams {
 
 export const buildParams = (params: UrlParams) => {
   let stringParams = '';
-  for(const key in params) {
+  for (const key in params) {
     stringParams = `${stringParams}&${key}=${params[key]}`;
   }
   return stringParams;
